вроде написал логику для сложных маршрутов, но пока не работает

+++ src/main/java/com/example/prac/data/res/TravelVariantDTO.java
@@ -7,10 +7,10 @@ import java.util.List;
 
 @Data
 public class TravelVariantDTO {
-    private double totalPrice;
+    private int totalPrice;
     private List<RouteDTO> routes;
 
-    public TravelVariantDTO(double totalPrice, List<RouteDTO> routes){
+    public TravelVariantDTO(int totalPrice, List<RouteDTO> routes){
         this.totalPrice = totalPrice;
         this.routes = routes;
     }
